requirements.txt
alembic==1.9.4
aniso8601==9.0.1
attrs==22.2.0
chardet==5.1.0
click==8.1.3
DataProperty==0.55.0
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mbstrdecoder==1.1.2
packaging==23.0
passlib==1.7.4
pathvalidate==2.5.2
pluggy==1.0.0
py==1.11.0
PyJWT==2.1.0
pytablewriter==0.64.2
pytest==7.2.1
pytest-html==3.2.0
pytest-md-report==0.3.0
pytest-metadata==2.0.4
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
tabledata==1.3.1
tcolorpy==0.1.2
tomli==2.0.1
typepy==1.3.0
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
